+++ packages/ragbits-core/CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## Unreleased
 
+- Allow custom source and element classes to be auto-imported (#343)
+- Make HuggingFaceSource row optional (#529)
+- Allow Source to be constructed from config (#529)
+
+## 0.16.0 (2025-04-29)
+
 ## 0.15.0 (2025-04-28)
 - Allow using sparse embeddings with Qdrant and local vector stores (#493)
 - Add support for sparse embeddings in the Pgvector Vector Store (#493)
 - Added secret key environment variable / generation for signatures across ragbits packages (#537)
-
 - Fix source interface definition (#535)
 
 ## 0.14.0 (2025-04-22)