+++ packages/ragbits-evaluate/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+- Add tests for ragbits-evaluate package (#390)
+- Integrate sources with dataloaders (#529)
+
+## 0.16.0 (2025-04-29)
+
+### Changed
+
+- ragbits-core updated to version v0.16.0
+
 ## 0.15.0 (2025-04-28)
 
 ### Changed